+++ Source/ACE.Common/GameConfiguration.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -27,19 +27,5 @@ namespace ACE.Common
         [System.ComponentModel.DefaultValue(60)]
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Populate)]
         public uint ShutdownInterval { get; set; }
-
-        /// <summary>
-        /// whether or not this server requires secure authentication.  for backwards compatibility, this is false
-        /// by default.
-        /// </summary>
-        [System.ComponentModel.DefaultValue(false)]
-        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Populate)]
-        public bool SecureAuthentication { get; set; }
-
-        /// <summary>
-        /// list of LoginServers that this server will accept tokens from.  If you are running
-        /// your own Login Server, you will need to make sure it is in this list.
-        /// </summary>
-        public List<string> AllowedAuthServers { get; set; } = new List<string>();
     }
 }